requirements.txt
django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
django-cors-headers==4.2.0
django-filter==23.5
django-tenant-schemas==1.11.0
django-guardian==2.4.0
celery==5.3.4
redis==5.0.1
boto3==1.29.7
requests==2.31.0
python-decouple==3.8
django-axes>=6.3.0
gunicorn==21.2.0
